Add ForecastService to make api requests
Created ForecastService that is responsible for constructing an
Alamofire request object and making the request. The request is
processed and uses ForecastResponseFactory to return an array of
FactoryResponse objects from the response JSON.

Added additional errors to handle request error and cases where the
response value is nil.

Constants defines (so far) values for the OWM request such as the
location we're querying and app id etc